// packages/frontend/src/shared/types/api.types.ts
export interface TradingSignal {
  id: number;
  createdAt: string;
  updatedAt: string;
  tokenAddress: string;
  symbol: string | null;
  signalType: string;
  activationReason: string | null;
  signalStrength: number;
  source: string | null;
}